OnewindRad working

+++ include/Init.cuh
@@ -1,2 +1,2 @@
-__host__ void Initialization (float *Dens, float *Vrad, float *Vtheta, float *Energy, float *Label, PlanetarySystem *sys);
-__host__ void InitLabel (float *Label, PlanetarySystem *sys);
+__host__ void Initialization (double *Dens, double *Vrad, double *Vtheta, double *Energy, double *Label, PlanetarySystem *sys);
+__host__ void InitLabel (double *Label, PlanetarySystem *sys);
