+++ include/Main.cuh
@@ -30,11 +30,11 @@
 
 using namespace std;
 
-__host__ void UpdateVelocitiesWithViscosity (float *VradInt, float *VthetaInt, float *dens, float DeltaT);
-__host__ void MultiplyPolarGridbyConstant (float *Dens);
+__host__ void UpdateVelocitiesWithViscosity (double *VradInt, double *VthetaInt, double *dens, double DeltaT);
+__host__ void MultiplyPolarGridbyConstant (double *Dens);
 __host__ void FreeCuda ();
-__host__ void FreeArrays (float *Dens, float *Vrad, float *Vtheta, float *energy, float *label);
-__host__ void DeviceToHostcudaMemcpy (float *Dens, float *energy, float *label, float *Temperature, float *Vrad, float *Vtheta);
+__host__ void FreeArrays (double *Dens, double *Vrad, double *Vtheta, double *energy, double *label);
+__host__ void DeviceToHostcudaMemcpy (double *Dens, double *energy, double *label, double *Temperature, double *Vrad, double *Vtheta);
 __host__ void Viscouscudamalloc ();
 __host__ void CreateArrays ();
-__host__ void Cudamalloc (float *label, float *Dens, float *Vrad, float *Vtheta);
+__host__ void Cudamalloc (double *label, double *Dens, double *Vrad, double *Vtheta);
