+++ src/Force.cu
@@ -2,13 +2,13 @@
 
 extern string OUTPUTDIR;
 
-extern float ROCHESMOOTHING, THICKNESSSMOOTHING, FLARINGINDEX;
+extern double ROCHESMOOTHING, THICKNESSSMOOTHING, FLARINGINDEX;
 
-extern float *CellAbscissa, *CellOrdinate, *Surf, *forcesxi, *forcesyi, *forcesxo;
-extern float *forcesyo;
+extern double *CellAbscissa, *CellOrdinate, *Surf, *forcesxi, *forcesyi, *forcesxo;
+extern double *forcesyo;
 
-extern float *Dens_d, *CellAbscissa_d, *CellOrdinate_d, *Surf_d;
-extern float *fxi_d, *fxo_d, *fyi_d, *fyo_d;
+extern double *Dens_d, *CellAbscissa_d, *CellOrdinate_d, *Surf_d;
+extern double *fxi_d, *fxo_d, *fyi_d, *fyo_d;
 
 extern double *Rmed, *Rmed_d;
 
@@ -16,12 +16,12 @@ extern int RocheSmoothing, size_grid, NRAD, NSEC, SelfGravity;
 
 extern dim3 dimGrid2, dimBlock2;
 
-__host__ void UpdateLog (Force *force, PlanetarySystem *sys, float *Dens, float *Energy, int TimeStep,
-  float PhysicalTime, int dimfxy)
+__host__ void UpdateLog (Force *force, PlanetarySystem *sys, double *Dens, double *Energy, int TimeStep,
+  double PhysicalTime, int dimfxy)
 {
   FILE *out;
   double x, y, r, m, vx, vy, smoothing, a, rh;
-  float *globalforce;
+  double *globalforce;
   char filename[500];
   char filename2[500];
   int i, nb;
@@ -88,9 +88,9 @@ __host__ void UpdateLog (Force *force, PlanetarySystem *sys, float *Dens, float 
 __host__ Force *AllocateForce (int dimfxy)
 {
   Force *force;
-  float *globalforce;
+  double *globalforce;
   force = (Force *)malloc(sizeof(Force));
-  globalforce = (float *)malloc(sizeof(float)*4*dimfxy);
+  globalforce = (double *)malloc(sizeof(double)*4*dimfxy);
   for (int i = 0; i < 4*dimfxy; i++)
     globalforce[i] = 0.0;
   force->GlobalForce = globalforce;
@@ -99,18 +99,18 @@ __host__ Force *AllocateForce (int dimfxy)
 
 
 
-__host__ void ComputeForce (Force *force, float *Dens, float x, float y, float rsmoothing,
-  float mass, int dimfxy, float a, float rh)
+__host__ void ComputeForce (Force *force, double *Dens, double x, double y, double rsmoothing,
+  double mass, int dimfxy, double a, double rh)
 {
-  float *globalforce;
+  double *globalforce;
   int k;
 
   globalforce = force->GlobalForce;
 
-  gpuErrchk(cudaMemset(fxi_d, 0, NRAD*NSEC*sizeof(float)));
-  gpuErrchk(cudaMemset(fxo_d, 0, NRAD*NSEC*sizeof(float)));
-  gpuErrchk(cudaMemset(fyi_d, 0, NRAD*NSEC*sizeof(float)));
-  gpuErrchk(cudaMemset(fyo_d, 0, NRAD*NSEC*sizeof(float)));
+  gpuErrchk(cudaMemset(fxi_d, 0, NRAD*NSEC*sizeof(double)));
+  gpuErrchk(cudaMemset(fxo_d, 0, NRAD*NSEC*sizeof(double)));
+  gpuErrchk(cudaMemset(fyi_d, 0, NRAD*NSEC*sizeof(double)));
+  gpuErrchk(cudaMemset(fyo_d, 0, NRAD*NSEC*sizeof(double)));
 
   for (k = 0; k < dimfxy; k++) {
     ComputeForceKernel<<<dimGrid2, dimBlock2>>>(CellAbscissa_d, CellOrdinate_d, Surf_d, Dens_d, x, y, rsmoothing,
@@ -134,9 +134,9 @@ __host__ void ComputeForce (Force *force, float *Dens, float x, float y, float r
   force->GlobalForce = globalforce;
 }
 
-__host__ float Compute_smoothing(float r)
+__host__ double Compute_smoothing(double r)
 {
-  float smooth;
+  double smooth;
   smooth = THICKNESSSMOOTHING * AspectRatioHost(r) * pow(r, 1.0+FLARINGINDEX);
   return smooth;
 }
